packages/Co-DETR/requirements.txt
addict==2.4.0
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1707022139797/work/certifi
mmcv-full==1.5.0
numpy==1.21.6
opencv-python==4.11.0.86
packaging==24.0
Pillow==9.5.0
platformdirs==4.0.0
PyYAML==6.0.1
regex==2024.4.16
tomli==2.0.1
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.1.6
typing_extensions==4.7.1
wincertstore==0.2
yapf==0.43.0
